W1 on VP fixes
